+++ src/components/Modals/Snaps/Snaps.tsx
@@ -1,5 +1,5 @@
 import { Modal, ModalCloseButton, ModalContent, ModalOverlay, useToast } from '@chakra-ui/react'
-import { useEffect } from 'react'
+import { useCallback, useEffect } from 'react'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 import { useIsSnapInstalled } from 'hooks/useIsSnapInstalled/useIsSnapInstalled'
 import { useModal } from 'hooks/useModal/useModal'
@@ -23,6 +23,10 @@ export const Snaps: React.FC<SnapsModalProps> = ({ isRemoved }) => {
     }
   }, [close, isSnapInstalled, toast])
 
+  const handleClose = useCallback(() => {
+    close()
+  }, [close])
+
   if (!isSnapsEnabled) return null
 
   return (
@@ -30,7 +34,7 @@ export const Snaps: React.FC<SnapsModalProps> = ({ isRemoved }) => {
       <ModalOverlay />
       <ModalContent minW='450px'>
         <ModalCloseButton />
-        <SnapContent isRemoved={isRemoved} />
+        <SnapContent isRemoved={isRemoved} onClose={handleClose} />
       </ModalContent>
     </Modal>
   )
