+++ src/context/WalletProvider/MetaMask/components/MetaMaskMenu.tsx
@@ -1,18 +1,19 @@
 import { MenuDivider, MenuItem, Skeleton, Tag } from '@chakra-ui/react'
 import { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
-import { enableShapeShiftSnap } from 'utils/snaps'
 import { useIsSnapInstalled } from 'hooks/useIsSnapInstalled/useIsSnapInstalled'
+import { useModal } from 'hooks/useModal/useModal'
 
 export const MetaMaskMenu = () => {
   const isSnapInstalled = useIsSnapInstalled()
   const translate = useTranslate()
+  const snapModal = useModal('snaps')
 
   const handleClick = useCallback(() => {
     if (isSnapInstalled === false) {
-      enableShapeShiftSnap()
+      snapModal.open({})
     }
-  }, [isSnapInstalled])
+  }, [isSnapInstalled, snapModal])
 
   const renderSnapStatus = useMemo(() => {
     if (isSnapInstalled === true) {
